added forms to users

+++ Unit-01/users-messages-forms/app.py
@@ -1,11 +1,14 @@
 from flask import Flask, request, redirect, render_template, url_for
 from flask_sqlalchemy import SQLAlchemy
 from flask_modus import Modus
+from forms import UserForm, MessageForm, DeleteForm
+import os
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = "postgres://localhost/users-messages"
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 modus = Modus(app)
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
 db = SQLAlchemy(app)
 
 class User(db.Model):
@@ -40,34 +43,46 @@ def root():
 
 @app.route('/users', methods=["GET", "POST"])
 def index():
+  delete_form = DeleteForm()
   if request.method == "POST":
-    new_user = User(request.form['first_name'], request.form['last_name'])
-    db.session.add(new_user)
-    db.session.commit()
-    return redirect(url_for('index'))
-  return render_template('users/index.html', users=User.query.all())
+    form = UserForm(request.form)
+    if form.validate ():
+      new_user = User(request.form['first_name'], request.form['last_name'])
+      db.session.add(new_user)
+      db.session.commit()
+      return redirect(url_for('index'))
+    else:
+      return render_template('users/new.html', form=form)
+  return render_template('users/index.html', users=User.query.all(), delete_form=delete_form)
 
 @app.route('/users/new')
 def new():
-  return render_template('users/new.html')
+  user_form = UserForm()
+  return render_template('users/new.html', form=user_form)
 
 @app.route('/users/<int:id>/edit')
 def edit(id):
   found_user = User.query.get(id)
-  return render_template('users/edit.html', user=found_user)
+  user_form = UserForm(obj=found_user)
+  return render_template('users/edit.html', user=found_user, form=user_form)
 
 @app.route('/users/<int:id>', methods=["GET", "PATCH", "DELETE"])
 def show(id):
   found_user = User.query.get(id)
   if request.method == b"PATCH":
-    found_user.first_name = request.form['first_name']
-    found_user.last_name = request.form['last_name']
-    db.session.add(found_user)
-    db.session.commit()
-    return redirect(url_for('index'))
+    form = UserForm(request.form)
+    if form.validate():
+      found_user.first_name = form.first_name.data
+      found_user.last_name = form.last_name.data
+      db.session.add(found_user)
+      db.session.commit()
+      return redirect(url_for('index'))
+    return render_template('users/edit.html', user=found_user, form=form)
   if request.method == b"DELETE":
-    db.session.delete(found_user)
-    db.session.commit()
+    delete_form = DeleteForm(request.form)
+    if delete_form.validate():
+      db.session.delete(found_user)
+      db.session.commit()
     return redirect(url_for('index'))
   return render_template('users/show.html', user=found_user)
